+++ b_dynamodb_common_test/integration/conftest.py
@@ -6,7 +6,6 @@ from b_aws_testing_framework.tools.cdk_testing.testing_manager import TestingMan
 
 CDK_PATH = f'{os.path.dirname(os.path.abspath(__file__))}'
 MANAGER = TestingManager(Credentials(), CdkToolConfig(CDK_PATH, destroy_before_preparing=False))
-MANAGER.set_global_prefix(override=False)
 
 # Import all fixtures.
 # noinspection PyUnresolvedReferences